requirements.txt
# pip install -r requirements.txt
flask=2.2.2
flask_cors
datetime
PyJWT
psycopg2
python-decouple
flask-swagger-ui
pytest
flask_sqlalchemy
flask_migrate